added portability service

+++ Backend/portability-request-service/src/main/java/com/example/mnp/model/Subscriber.java
@@ -9,7 +9,7 @@ public class Subscriber {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "subscriber_id")
-    private Integer subscriberId;
+    private Long subscriberId;   // Changed to Long (works better with PostgreSQL IDENTITY)
 
     @Column(name = "msisdn", nullable = false, unique = true)
     private Long msisdn;
@@ -33,20 +33,27 @@ public class Subscriber {
     private String serviceType;
 
     // Getters & Setters
-    public Integer getSubscriberId() { return subscriberId; }
-    public void setSubscriberId(Integer subscriberId) { this.subscriberId = subscriberId; }
+    public Long getSubscriberId() { return subscriberId; }
+    public void setSubscriberId(Long subscriberId) { this.subscriberId = subscriberId; }
+
     public Long getMsisdn() { return msisdn; }
     public void setMsisdn(Long msisdn) { this.msisdn = msisdn; }
+
     public Long getImsi() { return imsi; }
     public void setImsi(Long imsi) { this.imsi = imsi; }
+
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
+
     public String getIdType() { return idType; }
     public void setIdType(String idType) { this.idType = idType; }
+
     public String getIdNumber() { return idNumber; }
     public void setIdNumber(String idNumber) { this.idNumber = idNumber; }
+
     public Integer getCurrentProvider() { return currentProvider; }
     public void setCurrentProvider(Integer currentProvider) { this.currentProvider = currentProvider; }
+
     public String getServiceType() { return serviceType; }
     public void setServiceType(String serviceType) { this.serviceType = serviceType; }
 }
