+++ Backend/portability-request-service/src/main/java/com/example/mnp/repository/PortabilityRequestRepository.java
@@ -5,6 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PortabilityRequestRepository extends JpaRepository<PortabilityRequest, Integer> {
-    // 👈 The ID type must match entity @Id type
+public interface PortabilityRequestRepository extends JpaRepository<PortabilityRequest, Long> {
 }
